Reworked damage calculations especially for auto-attack, physical skills and dual wield weapons. Changed order in which damage/defence is calculated. Fixed: displayed damage in player profile, additional hit counts were not increasing damage, power was applied to skill damage, some effects could never be resisted, skill accuracy was never added to magical skills, calculations for evasion, parry and block and calculations for blockrate. This should fix: #BA2905 https://beyond-aion.com/forum/thread/2905-falscher-schaden-bei-geblockten-und-parierten-kritischen-treffern/ #BA2907 https://beyond-aion.com/forum/thread/2907-kantor-bergrutsch-keine-pr%C3%A4zision-500/ #BA2927 https://beyond-aion.com/forum/thread/2927-kantor-fehlende-machtscherben-funktion-bei-skills/ #BA2928 https://beyond-aion.com/forum/thread/2928-barde-gunner-%C3%A4thertech-verbraucht-machtscherben-bei-skills/ #BA2937 https://beyond-aion.com/forum/thread/2937-aetherwisting-bug-or-magical-debuff-applied-like-physical/ #BA2954 https://beyond-aion.com/forum/thread/2954-knockdown-buggy/ #BA2720 https://beyond-aion.com/forum/thread/2720-kantor-stab-hat-weniger-angriffsbuff-als-der-kolben/ #BA3024 https://beyond-aion.com/forum/thread/3024-fortgeschrittener-zweihandkampf/?postID=13077

+++ game-server/data/handlers/ai/OneDmgAI.java
@@ -17,12 +17,12 @@ public class OneDmgAI extends AggressiveNpcAI {
 	}
 
 	@Override
-	public int modifyDamage(Creature attacker, int damage, Effect effect) {
+	public float modifyDamage(Creature attacker, float damage, Effect effect) {
 		return 1;
 	}
 
 	@Override
-	public int modifyOwnerDamage(int damage, Creature effected, Effect effect) {
+	public float modifyOwnerDamage(float damage, Creature effected, Effect effect) {
 		return 1;
 	}
