+++ game-server/data/handlers/ai/instance/danuarSanctuary/AncientDanuarCoffinAI.java
@@ -24,7 +24,7 @@ public class AncientDanuarCoffinAI extends GeneralNpcAI {
 	}
 
 	@Override
-	public int modifyDamage(Creature attacker, int damage, Effect effect) {
+	public float modifyDamage(Creature attacker, float damage, Effect effect) {
 		return 1;
 	}
