+++ game-server/data/handlers/ai/instance/danuarSanctuary/SiegeDrill.java
@@ -36,7 +36,7 @@ public class SiegeDrill extends NpcAI {
 	}
 
 	@Override
-	public int modifyDamage(Creature attacker, int damage, Effect effect) {
+	public float modifyDamage(Creature attacker, float damage, Effect effect) {
 		return 0;
 	}
