+++ game-server/data/handlers/ai/instance/raksang/SealGeneratorAI.java
@@ -53,7 +53,7 @@ public class SealGeneratorAI extends AggressiveNpcAI {
 	}
 
 	@Override
-	public int modifyDamage(Creature attacker, int damage, Effect effect) {
+	public float modifyDamage(Creature attacker, float damage, Effect effect) {
 		return 1;
 	}
 }
