+++ game-server/data/handlers/ai/instance/shugoImperialTomb/LediarAssistantAI.java
@@ -21,7 +21,7 @@ public class LediarAssistantAI extends AggressiveNpcAI {
 	}
 
 	@Override
-	public int modifyOwnerDamage(int damage, Creature effected, Effect effect) {
+	public float modifyOwnerDamage(float damage, Creature effected, Effect effect) {
 		return 1;
 	}
