+++ game-server/data/handlers/ai/instance/shugoImperialTomb/TombAttackerAI.java
@@ -32,7 +32,7 @@ public class TombAttackerAI extends AggressiveNpcAI {
 	}
 
 	@Override
-	public int modifyOwnerDamage(int damage, Creature effected, Effect effect) {
+	public float modifyOwnerDamage(float damage, Creature effected, Effect effect) {
 		return 1;
 	}
