+++ game-server/data/handlers/ai/quests/AscensationNpcAI.java
@@ -18,7 +18,7 @@ public class AscensationNpcAI extends AggressiveNpcAI {
 	}
 
 	@Override
-	public int modifyOwnerDamage(int damage, Creature effected, Effect effect) {
+	public float modifyOwnerDamage(float damage, Creature effected, Effect effect) {
 		return 1;
 	}
