+++ game-server/data/handlers/ai/worlds/heiron/KlawspawnAI.java
@@ -37,7 +37,7 @@ public class KlawspawnAI extends GeneralNpcAI {
 	}
 
 	@Override
-	public int modifyDamage(Creature attacker, int damage, Effect effect) {
+	public float modifyDamage(Creature attacker, float damage, Effect effect) {
 		return 1;
 	}
