+++ game-server/src/com/aionemu/gameserver/model/stats/calc/ReverseStat.java
@@ -8,16 +8,16 @@ import com.aionemu.gameserver.model.stats.container.StatEnum;
  */
 public class ReverseStat extends Stat2 {
 
-	public ReverseStat(StatEnum stat, int base, Creature owner) {
+	public ReverseStat(StatEnum stat, float base, Creature owner) {
 		super(stat, base, owner);
 	}
 
-	public ReverseStat(StatEnum stat, int base, Creature owner, float bonusRate) {
+	public ReverseStat(StatEnum stat, float base, Creature owner, float bonusRate) {
 		super(stat, base, owner, bonusRate);
 	}
 
 	@Override
-	public void addToBase(int base) {
+	public void addToBase(float base) {
 		this.base -= base;
 		if (this.base < 0) {
 			this.base = 0;
@@ -25,7 +25,7 @@ public class ReverseStat extends Stat2 {
 	}
 
 	@Override
-	public void addToBonus(int bonus) {
+	public void addToBonus(float bonus) {
 		this.bonus -= bonusRate * bonus;
 	}
